// src/state/themeFamily.d.ts
export type AppearanceContent = 'light' | 'dark';
export type FontStyleContent = 'normal' | 'italic' | 'oblique';

/**
 * The content of a serialized theme family.
 */
export interface ThemeFamilyContent {
  author: string;
  name: string;
  themes: ThemeContent[];
}
/**
 * The content of a serialized theme.
 */
export interface ThemeContent {
  appearance: AppearanceContent;
  name: string;
  style: ThemeStyleContent;
}